requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
pillow==10.1.0
pytesseract==0.3.10
aiofiles==23.2.1
pydantic==2.5.2
pydantic-settings==2.1.0
structlog==23.2.0
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
psutil==5.9.6

